requirements.txt
aio5
Pillow~=9.2.0
streamlit~=1.16.0